+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Version](http://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 ### Added
+* Task `cache:clear:magento`
 
 ### Changed
 * Add `var` dir in `shared_dirs`. It makes more sense for multi-servers setup to have `var` dir shared among servers.
+* Change task name `clear:cache` to `cache:clear`
 
 ### Removed
 * Remove `var/.ip.maintenance` and `var/backups from shared configuration